+++ pkg/secretless/config/v2/config_test.go
@@ -101,7 +101,7 @@ func RunNewConfigTestCases(t *testing.T, label string, sampleContents string) {
 
 		assert.Equal(t, "postgres-db", cfg.Services[1].Name)
 		assert.Equal(t, "pg", cfg.Services[1].Connector)
-		assert.Equal(t, "tcp://0.0.0.0:5432", cfg.Services[1].ListenOn)
+		assert.Equal(t, NetworkAddress("tcp://0.0.0.0:5432"), cfg.Services[1].ListenOn)
 	})
 
 	t.Run(label + ": config hydration", func(t *testing.T) {
@@ -124,7 +124,7 @@ func RunNewConfigTestCases(t *testing.T, label string, sampleContents string) {
 
 		assert.Equal(t, "aws-proxy", cfg.Services[0].Name)
 		assert.Equal(t, "aws", cfg.Services[0].Connector)
-		assert.Equal(t, "tcp://0.0.0.0:8080", cfg.Services[0].ListenOn)
+		assert.Equal(t, NetworkAddress("tcp://0.0.0.0:8080"), cfg.Services[0].ListenOn)
 	})
 
 	t.Run(label + ": credential hydration", func(t *testing.T) {
